+++ RedPandaIDE/HighlighterManager.cpp
@@ -5,6 +5,7 @@
 #include "qsynedit/highlighter/asm.h"
 #include "qsynedit/Constants.h"
 #include "colorscheme.h"
+#include <QDebug>
 
 HighlighterManager highlighterManager;